+++ src/Core/CustomersManagement.Application/Features/Tours/Commands/UpdateTour/UpdateTourCommand.cs
@@ -8,8 +8,8 @@ public class UpdateTourCommand : IRequest<Unit>
     public string Name { get; set; }
     public DateTime TimeStart { get; set; }
     public DateTime TimeEnd { get; set; }
-    public int EntireCost { get; set; }
-    public int AdvancePaymentCost { get; set; }
+    public double EntireCost { get; set; }
+    public double AdvancePaymentCost { get; set; }
     public DateTime EntireAmountPaymentDeadline { get; set; }
     public DateTime AdvancePaymentDeadline { get; set; }
     public string Status { get; set; }
